改bug未选中无法删除

+++ material/src/views/Cost.vue
@@ -265,6 +265,7 @@ export default {
     },
     handleDelete () {
       console.log(this.currentRow)
+      if (JSON.stringify(this.currentRow) === '{}') return
       this.$confirm('此操作将永久删除该条数据, 是否继续?', '提示', { confirmButtonText: '确定', cancelButtonText: '取消', type: 'warning' })
         .then(() => {
           this.deleteItem(this.currentRow)
